Further development in COS code, ready for use
