+++ src/hst/stis_analysis.py
@@ -14,71 +14,7 @@ from astropy.io import fits
 from astropy.stats import poisson_conf_interval
 from scipy.integrate import simpson
 
-__all__ = ["integrate_flux", "timetag_split"]
-
-
-# This function integrates the flux within a wavelength range for given arrays
-# for wavelength and flux
-def integrate_flux(wavelength_range, wavelength_list, flux_list, gross_list,
-                   net_list, exposure_time):
-    """
-
-    Parameters
-    ----------
-    wavelength_range
-    wavelength_list
-    flux_list
-    gross_list
-    net_list
-    exposure_time
-
-    Returns
-    -------
-    integrated_flux
-    integrated_error
-    """
-    # Since the pixels may not range exactly in the interval above,
-    # we will need to deal with fractional pixels. But first, let's
-    # integrate the pixels that are fully inside the range
-    full_indexes = np.where((wavelength_list > wavelength_range[0]) & (
-                wavelength_list < wavelength_range[1]))[0]
-    full_pixel_flux = simpson(flux_list[full_indexes],
-                              wavelength_list[full_indexes])
-
-    # And now we deal with the flux in the fractional pixels
-    index_left = full_indexes[0]
-    index_right = full_indexes[-1]
-    pixel_width_left = wavelength_list[index_left] - wavelength_list[
-        index_left - 1]
-    fraction_left = 1 - (wavelength_range[0] - wavelength_list[
-        index_left - 1]) / pixel_width_left
-    pixel_width_right = wavelength_list[index_right + 1] - wavelength_list[
-        index_right]
-    fraction_right = 1 - (wavelength_list[index_right + 1] - wavelength_range[
-        1]) / pixel_width_right
-    fractional_flux_left = flux_list[index_left - 1] * fraction_left
-    fractional_flux_right = flux_list[index_right + 1] * fraction_right
-
-    # There is a bug in stistools that overestimates errors of time-tag split
-    # subexposures. So we will need to calculate them manually here
-    # based on the raw counts in the extracted spectra. Here we go.
-    full_pixel_gross = np.sum(gross_list[full_indexes])
-    fractional_gross_left = gross_list[index_left - 1] * fraction_left
-    fractional_gross_right = gross_list[index_right + 1] * fraction_right
-    integrated_gross = (full_pixel_gross + fractional_gross_left +
-                        fractional_gross_right)
-    sensitivity = flux_list[full_indexes] / net_list[full_indexes]
-    mean_sensitivity = np.mean(sensitivity)
-    gross_error = (poisson_conf_interval(integrated_gross,
-                                        interval='sherpagehrels') -
-                   integrated_gross)
-    integrated_error = (-gross_error[0] + gross_error[
-        1]) / 2 * mean_sensitivity / exposure_time  # Take the average error for
-    # simplicity
-
-    integrated_flux = (full_pixel_flux + fractional_flux_left +
-                       fractional_flux_right)
-    return integrated_flux, integrated_error
+__all__ = ["timetag_split", ]
 
 
 # Divide exposures into sub-exposures for TIME-TAG data and process them
